requirements.txt
Django==4.2.*
psycopg2-binary==2.9.*
django-environ==0.11.*
whitenoise==6.*
django-htmx==1.*
pytest==8.*
pytest-django==4.*
coverage==7.*
gunicorn==21.*
pytest-cov==5.*
